refactor(mobile): decompose inbox + my-tasks into feature modules
Both screens defined their row component, list chrome and empty/loading states
inline. Move the rows to the modules that own them (MentionRow -> messages/,
MyTaskRow -> tasks/) and lift the shared skeleton into shared/: CardList for the
FlatList chrome, ScreenStates for the centred loading and empty blocks.

inbox 282->110, my-tasks 259->110.

CardList owns the safe-area bottom inset because forgetting it puts the last
card under the gesture bar — invisible on a simulator, obvious on a device.

Scoped to these two screens on purpose. The projects list and build-unit index
have visually similar states built from DIFFERENT values (16px semibold empty
text vs 12px regular, 12px row gap vs 8, different centred padding), so folding
them in would have restyled them rather than deduplicated them. Both new files
carry that note.

Behavior-preserving. Typecheck holds at the 85-error baseline. Lint drops by one
pre-existing error: inbox's unescaped apostrophe was JSX text and is now a prop
value, so the rule no longer applies.
